requirements-dev.txt
numpy
pandas
scikit-learn==1.7.2
mlflow==2.22.0
dagshub
dvc
lightgbm
joblib
awscli
dvc_s3
fastapi
pydantic
python-jose
redis
pytest
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
psycopg2
argon2_cffi